Don't use the long type.
Instead of using the long type, use uint32_t. The reason for this is that
ntohl expects 32-bit parameters, and long on 64-bit machines is 64-bit.

The long type should be avoided unless needed to pass pointers, or if the
size of the type does not matter to the rest of the code.

+++ src/mediathread/demuxer/demuxer_sd.c
@@ -429,9 +429,9 @@ static int sd_read_packet(Resource * r)
         marker = (msg_buffer[1]>>7);
 
         if (tr->timestamp == 0)
-            tr->timestamp = ntohl(*(long*)(msg_buffer+4));
+            tr->timestamp = ntohl(*(uint32_t*)(msg_buffer+4));
 
-        delivery = (ntohl(*(long*)(msg_buffer+4)));
+        delivery = (ntohl(*(uint32_t*)(msg_buffer+4)));
 
         timestamp = delivery/((double)tr->properties->clock_rate);
         delivery = (delivery - tr->timestamp)
